separating tests into more descriptive file names. simplifying node/link comparisons
